+++ CD/Services/FileReaderWriter.cs
@@ -28,14 +28,14 @@ namespace CD.Services
             return block;
         }
 
-        public void Write(List<Tuple<int, double>> data, string path)
+        public void Write(List<Tuple<int, decimal>> data, string path)
         {
             using (StreamWriter writer = new StreamWriter(string.Concat(path, @"\13-out.txt")))
             {
                 writer.Write($"block size: {data.Count}");
                 writer.WriteLine();
 
-                foreach (Tuple<int, double> item in data)
+                foreach (Tuple<int, decimal> item in data)
                 {
                     writer.Write($"delta: {item.Item1} ");
                     writer.Write($"average: {item.Item2} ");
